efficiency, modal analysis
model is now aware of the defined components in the levels, and
gathering lists of components is now only done when necessary, since
this was repeatedly required and making those lists every time was a
waste of resources.

modal analysis can now output mode participation factors which can be
used to perform an actual modal response spectrum analysis.
effective modal mass calculations are incomplete.